+++ scopesim/optics/OLD_fov_manager.py
