+++ scopesim/tests/tests_effects/test_NonCommonPathAberation.py
@@ -31,6 +31,7 @@ def fov_Ks():
     _src = _single_table_source()
     _fov = FieldOfView(header=_fov_header(), waverange=(1.9, 2.4), area=1*u.m**2)
     _fov.extract_from(_src)
+    _fov.view()
     return _fov
