+++ scopesim/tests/tests_effects/test_Vibration.py
@@ -45,6 +45,7 @@ class TestInit:
 class TestApplyTo:
     def test_nothing_happens_if_apply_to_fov(self, fov_hdr):
         fov = FieldOfView(header=fov_hdr, waverange=[0.5, 2.5], area=1*u.m**2)
+        fov.view()
         fov.hdu.data = np.zeros((11, 11))
         fov.hdu.data[5, 5] = 1
         vibration = Vibration(**{"fwhm": 0.01, "pixel_scale": 0.004})
